requirements.txt
streamlit>=1.24.0
tensorflow-cpu>=2.16.1,<2.17.0
numpy>=1.24.0
opencv-python>=4.7.0
